SENTRY-1344: Move AuthorizationComponent from sentry-provider-common to sentry-core-common(Colin Ma, reviewed by Dapeng Sun)
